refactor edit properties dialog to separate class
